refactor: rename SearchParameters type to SearchRecord and move it along with search-string builder and extractor to PaginationPolitics static class
